+++ tests/testthat/test-query.R
@@ -1,0 +1,9 @@
+testthat::test_that("queries work", {
+  
+  testthat::skip_on_cran()
+  
+  q <- eurlex::elx_make_query("directive")
+  
+  testthat::expect_gt(nchar(q), 800)
+  
+})