+++ NEWS.md
@@ -1,3 +1,14 @@
+# eurlex 0.4.0
+
+## Major changes
+
+- download XML notices associated with Cellar URLs with `elx_download_xml()`
+- retrieve European Case Law Identifier (ECLI) with `elx_make_query(include_ecli = TRUE)`
+
+## Minor changes
+
+- host of smaller code improvements in `elx_fetch_data()`
+
 # eurlex 0.3.6
 
 ## Major changes